Initial refactor of stocks tests
